split on two componets
