+++ src/client/src/pages/PageNewProject/ProjectOutput/messages.ts
@@ -1,0 +1,9 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  outputPathTitle: {
+    id: "projectName.outputPathTitle",
+    defaultMessage: "Save To"
+  }
+});
+export default messages;
