+++ src/client/src/translations/whitelist_en.json
@@ -1,4 +1,8 @@
 [
+  "footer.navAriaLabel",
+  "footer.back",
+  "footer.next",
+  "footer.generate",
   "LocationSelection.title",
   "LocationSelection.subtitle",
   "LocationSelection.ariaDropdownLabel",
@@ -10,6 +14,34 @@
   "ResourceGroupSelection.ariaDropdownLabel",
   "ResourceGroupSelection.newResourceGroupLink",
   "ResourceGroupSelection.createNewResourceGroup",
+  "about.templatesVersion",
+  "about.wizardVersion",
+  "about.reportAnIssue",
+  "about.visitRepo",
+  "about.redirectLinkLabel",
+  "sortablePageList.hide",
+  "sortablePageList.show",
+  "sortablePageList.pages",
+  "draggableSidebarItem.changeItemName",
+  "draggableSidebarItem.pageNameMaxLength",
+  "draggableSidebarItem.deleteItem",
+  "AppServiceSelection.title",
+  "CosmosDbSelection.title",
+  "servicesList.SidebarItem.deleteItem",
+  "rightSidebar.services",
+  "rightSidebar.yourProjectDetails",
+  "rightSidebar.projectName",
+  "rightSidebar.location",
+  "rightSidebar.projectType",
+  "rightSidebar.frontendFramework",
+  "rightSidebar.backendFramework",
+  "rightSidebar.viewLicenses",
+  "about.reportAnIssueLabel",
+  "rightSidebar.redirectLinkLabel",
+  "rightSidebar.showAriaLabel",
+  "rightSidebar.hideAriaLabel",
+  "rightSidebar.showIcon",
+  "rightSidebar.hideIcon",
   "SubscriptionSelection.title",
   "SubscriptionSelection.subtitle",
   "SubscriptionSelection.ariaDropdownLabel",
@@ -47,10 +79,6 @@
   "CosmosModal.saveMessage",
   "CosmosModal.showAdvancedMode",
   "CosmosModal.hideAdvancedMode",
-  "footer.navAriaLabel",
-  "footer.back",
-  "footer.next",
-  "footer.generate",
   "generationItem.showLog",
   "generationItem.view",
   "generationItem.generationInProgress",
@@ -64,6 +92,8 @@
   "generationModal.reportAndIssue",
   "generationModal.createNewProject",
   "generationModal.openProject",
+  "viewLicensesModal.closeModalLabel",
+  "licenses.licenses",
   "pageAddPages.pagesTitleQuestion",
   "pageAddPages.iconAltMessage",
   "pageCard.details",
@@ -95,36 +125,6 @@
   "frameworkCard.details",
   "pageSelectFramework.frontendTitle",
   "pageSelectFramework.backendTitle",
-  "about.templatesVersion",
-  "about.wizardVersion",
-  "about.reportAnIssue",
-  "about.visitRepo",
-  "about.redirectLinkLabel",
-  "sortablePageList.hide",
-  "sortablePageList.show",
-  "sortablePageList.pages",
-  "draggableSidebarItem.changeItemName",
-  "draggableSidebarItem.pageNameMaxLength",
-  "draggableSidebarItem.deleteItem",
-  "AppServiceSelection.title",
-  "CosmosDbSelection.title",
-  "servicesList.SidebarItem.deleteItem",
-  "rightSidebar.services",
-  "rightSidebar.yourProjectDetails",
-  "rightSidebar.projectName",
-  "rightSidebar.location",
-  "rightSidebar.projectType",
-  "rightSidebar.frontendFramework",
-  "rightSidebar.backendFramework",
-  "rightSidebar.viewLicenses",
-  "about.reportAnIssueLabel",
-  "rightSidebar.redirectLinkLabel",
-  "rightSidebar.showAriaLabel",
-  "rightSidebar.hideAriaLabel",
-  "rightSidebar.showIcon",
-  "rightSidebar.hideIcon",
-  "viewLicensesModal.closeModalLabel",
-  "licenses.licenses",
   "servicesSelector.serviceGroupHostingName",
   "servicesSelector.serviceGroupHostingDescription",
   "servicesSelector.serviceGroupStorageName",
